Refs #37377 moved ivl enrollments query into queries namespace
